% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tests.R
\docType{data}
\name{pA.toy2}
\alias{pA.toy2}
\title{pA.toy2}
\format{
An object of class \code{data.frame}
}
\usage{
pA.toy2
}
\description{
Toy transcript expression data
}
\details{
Transcript expression data for testing the functions
}
\examples{
data <- APAlog::pA.toy2
}
\keyword{datasets}
