% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tests.R
\docType{data}
\name{pA_design}
\alias{pA_design}
\title{pA_design}
\format{
An object of class \code{data.frame}
}
\usage{
pA_design
}
\description{
Toy experimental design
}
\details{
experimental design for testing the functions
}
\examples{
data <- Apalog:pA_design
}
\keyword{datasets}
