% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tests.R
\name{pA_logit_1tr_pairs}
\alias{pA_logit_1tr_pairs}
\title{pA_logit_1tr_pairs}
\usage{
pA_logit_1tr_pairs(xdd, model)
}
\arguments{
\item{xdd}{input data including read counts of all pA sites of one transcript in all samples}

\item{model}{Regression model}
}
\value{
A dataframe containing log ratios (logistic regression coefficients), SE, z and p-values plus the names of
ref and alt pA sites.
}
\description{
Function to compare the usage of all pairs of pA sites of one transcript across samples
}
\details{
This function is used inside the pA_logit_pairwise function.
}
