% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tests.R
\name{pA_logit_dev}
\alias{pA_logit_dev}
\title{pA_logit_dev}
\usage{
pA_logit_dev(data, model, design = NULL, sample_ID = NULL)
}
\arguments{
\item{data}{Dataset containing poly A (pA) site read counts. This dataset must have a long shape, meaning that there should be only one
column containing read counts (and it MUST be named "count"). The first four columns must be called "transcript", "pA.site", "sample" and "count".
Thus, each row in \code{data} contains the read count for one pA - transcript - sample combination.
Other sample attributes beyond sample ID may be recorded in additional variables in this dataset, or provided separately through a design matrix
and a key variable (e.g. sample ID) connecting the \code{data} and \code{design} matrices.}

\item{model}{Regression model describing the dependence of pA site usage on sample attribute(s).}

\item{design}{(optional) Design matrix. A matrix describing sample attributes which can be used as predictors in the regression model.}

\item{sample_ID}{(optional) A key variable connecting the counts dataset (\code{data}) and the design matrix.}
}
\value{
Deviance test p-values (one per transcript).
}
\description{
Function to evaluate the overall effect of predictors on pA usage per transcript through a deviance test.
}
\details{
A deviance test compares the likelihood of the fitted model with its corresponding null. In other words, it tests how much the prediction of response
is improved by including the covariates, compared to a model with no covariates (the intercept only or null model). In the case of poly A site usage,
a logistic regression model is run first (but the outcome is not reported); then, instead of reporting the effects of individual predictors (covariates)
on the ratio of specific pairs of pA sites, a deviance test reveals the overall
relevance or informativeness of all the predictors in the model towards the pA site usage pattern for each transcript across samples.
}
\examples{
fit3_pA <- pA_logit_dev(pA.toy2, pA.site ~ cell_line, pA_design, "sample")
}
