% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tests.R
\name{pA_multi_logit}
\alias{pA_multi_logit}
\title{pA_multi_logit}
\usage{
pA_multi_logit(data, model, design = NULL, sample_ID = NULL,
  long_output = FALSE)
}
\arguments{
\item{data}{Dataset containing poly A (pA) site read counts. This dataset must have a long shape, meaning that there should be only one
column containing read counts (and it MUST be named "count"). The first four columns must be called "transcript", "pA.site", "sample" and "count".
Thus, each row in \code{data} contains the read count for one pA - transcript - sample combination.
Other sample attributes beyond sample ID may be recorded in additional variables in this dataset, or provided separately through a design matrix
and a key variable (e.g. sample ID) connecting the \code{data} and \code{design} matrices.}

\item{model}{Regression model describing the dependence of pA site usage on sample attribute(s).}

\item{design}{(optional) Design matrix. A matrix describing sample attributes which can be used as predictors in the regression model.}

\item{sample_ID}{(optional) A key variable connecting the counts dataset (\code{data}) and the design matrix.}

\item{long_output}{Logical variable describing output format. FALSE: Only regression coefficients and p-values are reported.
TRUE: Standard error of regression coefficients, and z scores are also included in the output. Default: FALSE.}
}
\value{
Log ratios (multinomial logistic regression coefficients) and p-values describing the effect of predictor(s) specified in the model on the usage ratio of each alternative to reference pA site
per trancript. If a long output is requested, SE and z scores are also reported.
}
\description{
Function to compare the usage of alternative poly A site(s) to a reference (canonical) site.
}
\details{
This function uses a multinomial logistic regression algorithm from the \emph{nnet} package. For each transcript, one poly A site (pA) is set as the
canonical (reference) site and the usage of alternative pA(s) is compared to this reference pA. By default, the pA that comes first alphabetically
is used as reference. The user can specify the reference pA for each transcript by adding a prefix like \code{0_} to its name.
If a transcript has n pA sites, n-1 comparisons will be made. Transcripts with only one pA site should be removed from \code{data}
before running this function.
}
\examples{
fit1_pA <- pA_multi_logit(pA.toy2, pA.site ~ cell_line, pA_design, "sample")
}
