% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tests.R
\name{remove_0_1_pA_transcripts}
\alias{remove_0_1_pA_transcripts}
\title{remove_0_1_pA_transcripts}
\usage{
remove_0_1_pA_transcripts(data)
}
\arguments{
\item{data}{Dataset containing poly A (pA) site read counts. This dataset must have a long shape, meaning that there should be only one
column containing read counts (and it MUST be named "count"). The first four columns must be called "transcript", "pA.site", "sample" and "count".
Thus, each row in \code{data} contains the read count for one pA - transcript - sample combination.
Other sample attributes beyond sample ID may be recorded in additional variables in this dataset, or provided separately through a design matrix
and a key variable (e.g. sample ID) connecting the \code{data} and \code{design} matrices.}
}
\value{
A subset of the input dataset where all transcripts are guaranteed to have two or more active pA sites.
}
\description{
Function to remove transcripts with fewer than two active pA sites from dataset
}
\details{
This function counts the number of pA sites with non-zero read counts for each transcripts and removes transripts
with fewer than two active pA sites. This is essential to avoid errors when running the regression models.
}
