% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tests.R
\name{adj_p}
\alias{adj_p}
\title{adj_p}
\usage{
adj_p(x, pcols, adj_method)
}
\arguments{
\item{x}{Dataframe containing a column or columns of p-values.}

\item{pcols}{A vector specifying unadjusted p-value columns in x}

\item{adj_method}{P-value adjustment method.
Options: "qvalue", "holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none".
"qvalue" calls the \emph{qvalue} package. Other methods are from base R.}
}
\value{
A data frame containing all the original information with the vector(s) of adjusted p-values
appended to the end.
}
\description{
Function to adjust p-values for multiple testing correction.
}
\details{
P-values e.g. coming from differential pA site usage on individual genes/transcripts
are correctd for multiple testing.
}
\examples{
fit3_pA_fdr <- adj_p(fit3_pA, pcols = 2, adj_method = "fdr")
}
