% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tests.R
\name{volcano_plot}
\alias{volcano_plot}
\title{volcano_plot}
\usage{
volcano_plot(
  fit,
  x,
  xlab = "Ln fold change",
  y,
  ylab = "-Log10 FDR",
  title = "LMCN data, metastatic vs non-metastatic",
  titleLabSize = 12,
  border = "full",
  pCutoff = 0.001,
  FCcutoff = 1.5,
  xlim = c(-5, 5),
  ylim = c(0, 10)
)
}
\arguments{
\item{fit}{Output dataframe of the pA_logit_dev or pA_multi_logit function}

\item{x}{Name of the column in the dataframe that contains the log fold change}

\item{y}{Name of the column in the dataframe that contains the corrected p-values}
}
\value{
A volcano plot visualisation
}
\description{
Function to make the volcano plot from the p-values matrix
}
\details{
A visualisation of the volcano plot resulting from the p-values and log fold change
}
\examples{
vplot <- volcano_plot(fit1, x='logfdr', y='p_values')
}
